Hook into Gunicorn to display launch messages
This means that we can not longer use the lazy monkey-patching
mechanism, but need to duplicate some code from the stock runserver
command. I guess that is okay.

We also now passes some information to the subprocess through
environ. I don't like it, but I guess that is inevitable. Django
does this, too, anyway.